modifications : created interface for LeaveService
